Edited dtos

+++ src/main/java/tetrad/virtualmuseum/repository/GalleryRepo.java
@@ -3,8 +3,8 @@ package tetrad.virtualmuseum.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tetrad.virtualmuseum.DAO.Gallery;
 
-public interface GalleryRepo extends JpaRepository<Gallery, Integer> {
+public interface GalleryRepo extends JpaRepository<Gallery, Long> {
 
 // findAll is a default, usable function. Can omit here.
-    Gallery findGalleryById(int id);
+    Gallery findGalleryById(Long id);
 }
